url_shortifier/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
hashids==1.3.1
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
